refactor: modularize data source handling
